+++ app/policies/location_policy.rb
@@ -1,0 +1,25 @@
+class LocationPolicy < ApplicationPolicy
+  def show?
+    user_is_current_user
+  end
+
+  def new?
+    user_is_current_user
+  end
+
+  def create?
+    user_is_current_user
+  end
+
+  def edit?
+    user_is_current_user
+  end
+
+  def update?
+    user_is_current_user
+  end
+
+  def destroy?
+    user_is_current_user
+  end
+end