+++ db/migrate/20180615124050_add_locations.rb
@@ -1,0 +1,8 @@
+class AddLocations < ActiveRecord::Migration
+  def change
+    create_table :locations do |t|
+      t.string  :name
+      t.integer :user_id
+    end
+  end
+end