+++ spec/factories/location.rb
@@ -1,0 +1,7 @@
+FactoryGirl.define do
+  factory :location do
+    user
+
+    name { Faker::Name.name }
+  end
+end