+++ spec/features/dm_log_entries/create_dm_log_entries_spec.rb
@@ -29,7 +29,7 @@ RSpec.feature 'DM Log Entries', type: :feature do
       fill_in 'Downtime +/-',       with: '111'
       fill_in 'Renown',             with: '44'
       fill_in 'Mission',            with: '55'
-      fill_in 'Location DMed',      with: 'Origins'
+      fill_in 'Location',           with: 'Origins'
       fill_in 'Notes',              with: 'Some Words'
 
       fill_in 'Date Assigned',      with: '' #Hack for calendar popout