+++ spec/features/dm_log_entries/edit_dm_log_entries_spec.rb
@@ -26,7 +26,7 @@ RSpec.feature 'DM Log Entries', type: :feature do
       fill_in 'Downtime +/-',       with: '111'
       fill_in 'Renown',             with: '44'
       fill_in 'Mission',            with: '55'
-      fill_in 'Location DMed',      with: 'Origins'
+      fill_in 'Location',           with: 'Origins'
       fill_in 'Notes',              with: 'Some Words'
 
       fill_in 'Date Assigned',      with: '' #Hack for calendar popout
@@ -89,7 +89,7 @@ RSpec.feature 'DM Log Entries', type: :feature do
         fill_in 'Downtime +/-',       with: '111'
         fill_in 'Renown',             with: '44'
         fill_in 'Mission',            with: '55'
-        fill_in 'Location DMed',      with: 'Origins'
+        fill_in 'Location',           with: 'Origins'
         fill_in 'Notes',              with: 'Some Words'
 
         fill_in 'Date Assigned',      with: '' #Hack for calendar popout