+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 v1.7.2
 ---------------
 - Add DM Log Entry search functionalities
+- Make Location a dropdown in log entry forms
 
 v1.7.1
 ---------------